Refactoring. Account was renamed to Settings
