+++ src/containers/app/AppWrapper/index.js
@@ -11,7 +11,7 @@ import KycAlertPopup from '../KycAlertPopup';
 import Dashboard from '../../dashboard/Dashboard';
 import Referrals from '../../referrals/Referrals';
 import Transactions from '../../transactions/Transactions';
-import Account from '../../account/Account';
+import Settings from '../../settings/Settings';
 import SendTokens from '../../sendTokens/SendTokens';
 import Verification from '../../../components/verification/Verification';
 import Error404 from '../../../components/common/Error404';
@@ -42,7 +42,7 @@ class AppWrapper extends Component {
           <Route exact path={namedRoutes.dashboard} component={Dashboard}/>
           <Route exact path={namedRoutes.referrals} component={Referrals}/>
           <Route exact path={namedRoutes.transactions} component={Transactions}/>
-          <Route exact path={namedRoutes.account} component={Account}/>
+          <Route exact path={namedRoutes.settings} component={Settings}/>
           <Route exact path={namedRoutes.sendTokens} component={SendTokens}/>
           <Route exact path={namedRoutes.verification} component={Verification}/>
           <Redirect exact from="/" to={namedRoutes.dashboard} />
