+++ src/containers/settings/Settings/index.js
@@ -2,18 +2,18 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import s from './styles.css';
 
-import { openEnableTwoFactorAuthPopup } from '../../../redux/modules/account/enableTwoFactorAuth';
-import { openDisableTwoFactorAuthPopup } from '../../../redux/modules/account/disableTwoFactorAuth';
+import { openEnableTwoFactorAuthPopup } from '../../../redux/modules/settings/enableTwoFactorAuth';
+import { openDisableTwoFactorAuthPopup } from '../../../redux/modules/settings/disableTwoFactorAuth';
 
 import Info from '../Info';
 import ChangePasswordPopup from '../ChangePasswordPopup';
 import VerifyChangePasswordPopup from '../VerifyChangePasswordPopup';
-import TwoFactorAuth from '../../../components/account/TwoFactorAuth';
-import Address from '../../../components/account/Address';
+import TwoFactorAuth from '../../../components/settings/TwoFactorAuth';
+import Address from '../../../components/settings/Address';
 import EnableTwoFactorAuthPopup from '../EnableTwoFactorAuthPopup';
 import DisableTwoFactorAuthPopup from '../DisableTwoFactorAuthPopup';
 
-class Account extends Component {
+class Settings extends Component {
   render() {
     const {
       openEnableTwoFactorAuthPopup,
@@ -59,4 +59,4 @@ export default connect(
     openEnableTwoFactorAuthPopup,
     openDisableTwoFactorAuthPopup
   }
-)(Account);
+)(Settings);
