+++ src/redux/rootReducer.js
@@ -15,9 +15,9 @@ import resetPassword from './modules/auth/resetPassword';
 import referrals from './modules/referrals/referrals';
 import invitePopup from './modules/referrals/invitePopup';
 
-import changePasswordReducer from './modules/account/changePassword';
-import enableTwoFactorAuth from './modules/account/enableTwoFactorAuth';
-import disableTwoFactorAuth from './modules/account/disableTwoFactorAuth';
+import changePasswordReducer from './modules/settings/changePassword';
+import enableTwoFactorAuth from './modules/settings/enableTwoFactorAuth';
+import disableTwoFactorAuth from './modules/settings/disableTwoFactorAuth';
 
 import dashboard from './modules/dashboard/dashboard';
 import buyTokens from './modules/dashboard/buyTokens';
